Moved user setting into a new feature
