+++ src/features/common/users/index.ts
@@ -1,0 +1,4 @@
+export { default as UserContactDataCard } from './components/UserContactDataCard.vue';
+export { default as UserAddressCard } from './components/UserAddressCard.vue';
+export { default as UserDangerZoneCard } from './components/UserDangerZoneCard.vue';
+export { default as AccountSettingsView } from './views/AccountSettingsView.vue';
