+++ src/pages/tenancies/account-settings.vue
@@ -6,7 +6,7 @@ meta:
 </route>
 
 <script setup lang="ts">
-import AccountSettingsView from "@/views/AccountSettingsView.vue";
+import { AccountSettingsView } from "@/features/common/users";
 </script>
 
 <template>
