+++ test/features/common/users/components/UserDangerZoneCard.spec.ts
@@ -1,6 +1,6 @@
 import { describe, test, expect, beforeEach, afterEach, vi } from 'vitest';
 import { mount, VueWrapper, flushPromises } from '@vue/test-utils';
-import UserDangerZoneCard from '@/components/UserDangerZoneCard.vue';
+import UserDangerZoneCard from '@/features/common/users/components/UserDangerZoneCard.vue';
 import { userService } from '@/services/UserService';
 
 // ---- Mocks ----
